requirements.txt
openai==1.3.0
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
requests==2.31.0
schedule==1.2.0
pydub==0.25.1
python-dotenv==1.0.0

